Create Configuration feature folder and move test classes
